Enable use of auditing with non-integer primary keys
The audit models assume that the primary keys of the models being
audited will be integers. This is not in general true. Move to storing
the string representation of the primary key instead of the integer
value.

Add a migration which alters the table. Postgres should be smart enough
to coerce the values.

Required to allow auditing the IAR.