middleware: improve compatibility with Django versions
We try to make our middleware compatible with old- and new-style
middleware. When used as old-style, the __init__ method is not passed
the "get_response" callable.

Work around this by defaulting get_response to None and asserting in the
new-style code paths that we were given an non-None get_response.